requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
pymupdf>=1.23.0
python-docx>=1.1.0
openpyxl>=3.1.0
pyyaml>=6.0

# NLP / PII detection
spacy>=3.5.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Storage / utilities
sqlalchemy>=2.0.0
pydantic>=2.0.0
